requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
scikit-learn>=1.3.0
pymysql>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.15.0
